Cleanup EFR32 Examples
