+++ examples/light-switch-app/efr32/src/main.cpp
@@ -17,63 +17,13 @@
  *    limitations under the License.
  */
 
-#include <stdio.h>
-#include <stdlib.h>
-
-#include <stdarg.h>
-#include <stdbool.h>
-#include <stdint.h>
-
-#include <FreeRTOS.h>
-#include <mbedtls/threading.h>
-
-#include <lib/support/CHIPMem.h>
-#include <lib/support/CHIPPlatformMemory.h>
-#include <platform/CHIPDeviceLayer.h>
-#include <platform/KeyValueStoreManager.h>
-
 #include <AppTask.h>
 
 #include "AppConfig.h"
 #include "init_efrPlatform.h"
 #include "sl_simple_button_instances.h"
 #include "sl_system_kernel.h"
-#include <app/server/Server.h>
-
-#ifdef HEAP_MONITORING
-#include "MemMonitoring.h"
-#endif
-
-#if DISPLAY_ENABLED
-#include "lcd.h"
-#endif
-
-#include <mbedtls/platform.h>
-#if CHIP_ENABLE_OPENTHREAD
-#include <inet/EndPointStateOpenThread.h>
-#include <openthread/cli.h>
-#include <openthread/dataset.h>
-#include <openthread/error.h>
-#include <openthread/heap.h>
-#include <openthread/icmp6.h>
-#include <openthread/instance.h>
-#include <openthread/link.h>
-#include <openthread/platform/openthread-system.h>
-#include <openthread/tasklet.h>
-#include <openthread/thread.h>
-#endif // CHIP_ENABLE_OPENTHREAD
-
-#if defined(RS911X_WIFI) || defined(WF200_WIFI)
-#include "wfx_host_events.h"
-#endif /* RS911X_WIFI */
-
-#if PW_RPC_ENABLED
-#include "Rpc.h"
-#endif
-
-#ifdef ENABLE_CHIP_SHELL
-#include "matter_shell.h"
-#endif
+#include <matter_config.h>
 
 #define BLE_DEV_NAME "SiLabs-Light-Switch"
 using namespace ::chip;
@@ -84,170 +34,25 @@ using namespace ::chip::DeviceLayer;
 
 volatile int apperror_cnt;
 // ================================================================================
-// App Error
-//=================================================================================
-void appError(int err)
-{
-    EFR32_LOG("!!!!!!!!!!!! App Critical Error: %d !!!!!!!!!!!", err);
-    portDISABLE_INTERRUPTS();
-    while (1)
-        ;
-}
-
-void appError(CHIP_ERROR error)
-{
-    appError(static_cast<int>(error.AsInteger()));
-}
-
-// ================================================================================
-// FreeRTOS Callbacks
-// ================================================================================
-extern "C" void vApplicationIdleHook(void)
-{
-    // FreeRTOS Idle callback
-
-    // Check CHIP Config nvm3 and repack flash if necessary.
-    Internal::EFR32Config::RepackNvm3Flash();
-}
-#if CHIP_ENABLE_OPENTHREAD
-// ================================================================================
-// Matter Networking Callbacks
-// ================================================================================
-void LockOpenThreadTask(void)
-{
-    chip::DeviceLayer::ThreadStackMgr().LockThreadStack();
-}
-
-void UnlockOpenThreadTask(void)
-{
-    chip::DeviceLayer::ThreadStackMgr().UnlockThreadStack();
-}
-#endif
-// ================================================================================
 // Main Code
 // ================================================================================
 int main(void)
 {
     init_efrPlatform();
-    mbedtls_platform_set_calloc_free(CHIPPlatformMemoryCalloc, CHIPPlatformMemoryFree);
-
-#if PW_RPC_ENABLED
-    chip::rpc::Init();
-#endif
-
-#ifdef HEAP_MONITORING
-    MemMonitoring::startHeapMonitoring();
-#endif
-
-    EFR32_LOG("==================================================");
-    EFR32_LOG("chip-efr32-light-switch-example starting");
-    EFR32_LOG("==================================================");
-
-    EFR32_LOG("Init CHIP Stack");
-    // Init Chip memory management before the stack
-    chip::Platform::MemoryInit();
-
-    CHIP_ERROR ret = PlatformMgr().InitChipStack();
-    if (ret != CHIP_NO_ERROR)
-    {
-        EFR32_LOG("PlatformMgr().InitChipStack() failed");
-        appError(ret);
-    }
-    chip::DeviceLayer::ConnectivityMgr().SetBLEDeviceName(BLE_DEV_NAME);
-#if CHIP_ENABLE_OPENTHREAD
-    EFR32_LOG("Initializing OpenThread stack");
-    ret = ThreadStackMgr().InitThreadStack();
-    if (ret != CHIP_NO_ERROR)
-    {
-        EFR32_LOG("ThreadStackMgr().InitThreadStack() failed");
-        appError(ret);
-    }
-
-#if CHIP_DEVICE_CONFIG_THREAD_FTD
-    ret = ConnectivityMgr().SetThreadDeviceType(ConnectivityManager::kThreadDeviceType_Router);
-#else // CHIP_DEVICE_CONFIG_THREAD_FTD
-#if CHIP_DEVICE_CONFIG_ENABLE_SED
-    ret = ConnectivityMgr().SetThreadDeviceType(ConnectivityManager::kThreadDeviceType_SleepyEndDevice);
-#else  // CHIP_DEVICE_CONFIG_ENABLE_SED
-    ret = ConnectivityMgr().SetThreadDeviceType(ConnectivityManager::kThreadDeviceType_MinimalEndDevice);
-#endif // CHIP_DEVICE_CONFIG_ENABLE_SED
-#endif // CHIP_DEVICE_CONFIG_THREAD_FTD
-    if (ret != CHIP_NO_ERROR)
-    {
-        EFR32_LOG("ConnectivityMgr().SetThreadDeviceType() failed");
-        appError(ret);
-    }
-#endif // CHIP_ENABLE_OPENTHREAD
-
-    chip::DeviceLayer::PlatformMgr().LockChipStack();
-    // Init ZCL Data Model
-    static chip::CommonCaseDeviceServerInitParams initParams;
-    (void) initParams.InitializeStaticResourcesBeforeServerInit();
-#if CHIP_ENABLE_OPENTHREAD
-    chip::Inet::EndPointStateOpenThread::OpenThreadEndpointInitParam nativeParams;
-    nativeParams.lockCb                = LockOpenThreadTask;
-    nativeParams.unlockCb              = UnlockOpenThreadTask;
-    nativeParams.openThreadInstancePtr = chip::DeviceLayer::ThreadStackMgrImpl().OTInstance();
-    initParams.endpointNativeParams    = static_cast<void *>(&nativeParams);
-#endif
-    chip::Server::GetInstance().Init(initParams);
-    chip::DeviceLayer::PlatformMgr().UnlockChipStack();
-
-    EFR32_LOG("Starting Platform Manager Event Loop");
-    ret = PlatformMgr().StartEventLoopTask();
-    if (ret != CHIP_NO_ERROR)
-    {
-        EFR32_LOG("PlatformMgr().StartEventLoopTask() failed");
-        appError(ret);
-    }
-
-#ifdef WF200_WIFI
-    // Start wfx bus communication task.
-    wfx_bus_start();
-#ifdef SL_WFX_USE_SECURE_LINK
-    wfx_securelink_task_start(); // start securelink key renegotiation task
-#endif                           // SL_WFX_USE_SECURE_LINK
-#endif                           /* WF200_WIFI */
-
-#if CHIP_ENABLE_OPENTHREAD
-    EFR32_LOG("Starting OpenThread task");
-
-    // Start OpenThread task
-    ret = ThreadStackMgrImpl().StartThreadTask();
-    if (ret != CHIP_NO_ERROR)
-    {
-        EFR32_LOG("ThreadStackMgr().StartThreadTask() failed");
-        appError(ret);
-    }
-#endif // CHIP_ENABLE_OPENTHREAD
-#ifdef RS911X_WIFI
-    /*
-     * Start up any RSI interface stuff
-     * (Not required) - Note that wfx_wifi_start will deal with
-     * starting up a rsi task - which will initialize the SPI interface.
-     */
-#endif
+    if (EFR32MatterConfig::InitMatter(BLE_DEV_NAME) != CHIP_NO_ERROR)
+        appError(CHIP_ERROR_INTERNAL);
 
     EFR32_LOG("Starting App Task");
-    ret = GetAppTask().StartAppTask();
-    if (ret != CHIP_NO_ERROR)
-    {
-        EFR32_LOG("GetAppTask().Init() failed");
-        appError(ret);
-    }
-
-#ifdef ENABLE_CHIP_SHELL
-    chip::startShellTask();
-#endif
+    if (GetAppTask().StartAppTask() != CHIP_NO_ERROR)
+        appError(CHIP_ERROR_INTERNAL);
 
     EFR32_LOG("Starting FreeRTOS scheduler");
     sl_system_kernel_start();
 
-    chip::Platform::MemoryShutdown();
-
     // Should never get here.
+    chip::Platform::MemoryShutdown();
     EFR32_LOG("vTaskStartScheduler() failed");
-    appError(ret);
+    appError(CHIP_ERROR_INTERNAL);
 }
 
 void sl_button_on_change(const sl_button_t * handle)
