+++ examples/ota-requestor-app/efr32/BUILD.gn
@@ -60,7 +60,7 @@ lcd_on = show_qr_code
 efr32_sdk("sdk") {
   sources = [
     "${efr32_project_dir}/include/CHIPProjectConfig.h",
-    "${efr32_project_dir}/include/FreeRTOSConfig.h",
+    "${examples_plat_dir}/FreeRTOSConfig.h",
   ]
 
   include_dirs = [
