+++ examples/pigweed-app/efr32/BUILD.gn
@@ -30,7 +30,7 @@ examples_plat_dir = "${chip_root}/examples/platform/efr32"
 efr32_sdk("sdk") {
   sources = [
     "${efr32_project_dir}/include/CHIPProjectConfig.h",
-    "${efr32_project_dir}/include/FreeRTOSConfig.h",
+    "${examples_plat_dir}/FreeRTOSConfig.h",
   ]
 
   include_dirs = [
