+++ examples/platform/efr32/efr32_utils.h
@@ -1,7 +1,7 @@
 /*
  *
  *    Copyright (c) 2020 Project CHIP Authors
- *    Copyright (c) 2019 Google LLC.
+ *    Copyright (c) 2022 Silabs.
  *    All rights reserved.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,14 +17,20 @@
  *    limitations under the License.
  */
 
-#pragma once
+// EFR Logging
+#ifdef __cplusplus
+extern "C" {
+#endif
 
-#include "efr32_utils.h"
+void efr32LogInit(void);
 
-// ---- Lighting Example App Config ----
+void efr32Log(const char * aFormat, ...);
+#define EFR32_LOG(...) efr32Log(__VA_ARGS__);
+void appError(int err);
 
-#define APP_TASK_NAME "Lit"
+#ifdef __cplusplus
+}
 
-// Time it takes in ms for the simulated actuator to move from one
-// state to another.
-#define ACTUATOR_MOVEMENT_PERIOS_MS 10
+#include <lib/core/CHIPError.h>
+void appError(CHIP_ERROR error);
+#endif
